pyproject.toml
[project]
name = "phi-deidentification"
version = "0.1.0"
description = "Autoresearch loop for PHI de-identification using GLiNER2"
requires-python = ">=3.10"
dependencies = [
    "gliner2>=1.2.0",
    "faker>=37.0.0",
    "urllib3>=2.6.3",
    "requests>=2.33.1",
    "pytesseract>=0.3.13",
    "pillow>=12.2.0",
    "pypdf>=6.10.0",
    "fastapi>=0.135.3",
    "uvicorn>=0.44.0",
    "python-multipart>=0.0.26",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
